% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederalistPapers.R
\docType{package}
\name{FederalistPapers}
\alias{FederalistPapers}
\alias{FederalistPapers-package}
\title{FederalistPapers}
\format{
A dataframe with 19,752 elements
\describe{
\item{text}{Line of text}
\item{linenumber}{Line number within paper}
\item{paper}{Paper number (1 through 86}
\item{roman_number}{Roman numeral equivalent of paper number}
}
}
\source{
\url{https://www.gutenberg.org/ebooks/18}
}
\usage{
FederalistPapers
}
\description{
This package contains the text of the Federalist Papers

A data set containing the text of the Federalist Papers
}
\keyword{datasets}
